# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference : TMF 641 - Service Ordering Management\n\n##\
    \ Release : 18.5 - Dec 2018\n\nThe Service Order API provides a standardized mechanism\
    \ for managing Service Order, a type of order which can be used to place an order\
    \ between a customer and a service provider or between a service provider and\
    \ a partner and vice versa.\n\n### Service Order resource\nA service order will\
    \ describe a list of service order items.  A service order item references an\
    \ action on an existing  or  future  service.  By  service  we  designed  Customer\
    \ Facing  Service  (CFS)  as  well  as  Resource Facing Service (RFS).\nFrom a\
    \ component perspective, a service order should be available\n- from a Service\
    \ Orchestration Component (and it could mix CFS and RFS)\n- from an Infrastructure\
    \ Control & Management component (and it would have only RFS)\n\n### Service Order\
    \ API performs the following operations on service order :\n- Retrieval of a service\
    \ order or a collection of service orders depending on filter criteria\n- Partial\
    \ update of a service order (including updating rules)\n- Creation of a service\
    \ order (including default values and creation rules)\n- Deletion of service order\
    \ (for administration purposes)\n- Notification of events on Service order\n\n\
    Copyright © TM Forum 2018. All Rights Reserved"
  version: "4.0.0"
...
host: "localhost:8080"
basePath: "/tmf-api"
...
schemes:
...
consumes:
...
produces:
...
paths:
  /serviceOrder:
    get:
      tags:
      - "serviceOrder"
      summary: "List or find ServiceOrder objects"
      description: "This operation list or find ServiceOrder entities"
      operationId: "listServiceOrder"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to be provided in response"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "Requested index for start of resources to be provided in response"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Requested number of resources to be provided in response"
        required: false
        type: "integer"
      responses:
        200:
          description: "Success"
          headers:
            X-Total-Count:
              type: "integer"
              description: "Total number of items matching criteria"
            X-Result-Count:
              type: "integer"
              description: "Actual number of items returned in the response body"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceOrder"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceOrder"
    post:
      tags:
      - "serviceOrder"
      summary: "Creates a ServiceOrder"
      description: "This operation creates a ServiceOrder entity."
      operationId: "createServiceOrder"
      parameters:
      - in: "body"
        name: "serviceOrder"
        description: "The ServiceOrder to be created"
        required: true
        schema:
          $ref: "#/definitions/ServiceOrder_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/ServiceOrder"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceOrder"
  /serviceOrder/{id}:
    get:
      tags:
      - "serviceOrder"
      summary: "Retrieves a ServiceOrder by ID"
      description: "This operation retrieves a ServiceOrder entity. Attribute selection\
        \ is enabled for all first level attributes."
      operationId: "retrieveServiceOrder"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceOrder"
        required: true
        type: "string"
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to provide in response"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/ServiceOrder"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceOrder"
    delete:
      tags:
      - "serviceOrder"
      summary: "Deletes a ServiceOrder"
      description: "This operation deletes a ServiceOrder entity."
      operationId: "deleteServiceOrder"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceOrder"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceOrder"
    patch:
      tags:
      - "serviceOrder"
      summary: "Updates partially a ServiceOrder"
      description: "This operation updates partially a ServiceOrder entity."
      operationId: "patchServiceOrder"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceOrder"
        required: true
        type: "string"
      - in: "body"
        name: "serviceOrder"
        description: "The ServiceOrder to be updated"
        required: true
        schema:
          $ref: "#/definitions/ServiceOrder_Update"
      responses:
        200:
          description: "Updated"
          schema:
            $ref: "#/definitions/ServiceOrder"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceOrder"
...
definitions:
...
  ServiceOrder:
    type: "object"
    required:
    - "orderItem"
    properties:
      id:
        type: "string"
        description: "ID created on repository side"
      href:
        type: "string"
        description: "Hyperlink to access the order"
      category:
        type: "string"
        description: "Used to categorize the order, useful for the OM system, such\
          \ as: Broadband, TVOption"
      completionDate:
        type: "string"
        format: "date-time"
        description: "Effective delivery date amended by the provider"
      description:
        type: "string"
        description: "A free-text description of the service order"
      expectedCompletionDate:
        type: "string"
        format: "date-time"
        description: "Expected delivery date amended by the provider"
      externalId:
        type: "string"
        description: "ID given by the consumer to facilitate searches"
      notificationContact:
        type: "string"
        description: "Contact attached to the order to send back information regarding\
          \ this order"
      orderDate:
        type: "string"
        format: "date-time"
      priority:
        type: "string"
        description: "Can be used by consumers to prioritize orders in a Service Order\
          \ Management system"
      requestedCompletionDate:
        type: "string"
        format: "date-time"
        description: "Requested delivery date from the requestors perspective"
      requestedStartDate:
        type: "string"
        format: "date-time"
        description: "Order start date wished by the requestor"
      startDate:
        type: "string"
        format: "date-time"
        description: "Date when the order was started for processing"
      note:
        type: "array"
        description: "Extra-information about the order; e.g. useful to add extra\
          \ delivery information that could be useful for a human process"
        items:
          $ref: "#/definitions/Note"
      orderItem:
        type: "array"
        description: "A list of service order items to be processed by this order"
        items:
          $ref: "#/definitions/ServiceOrderItem"
        minItems: 1
      orderRelationship:
        type: "array"
        description: "A list of service orders related to this order (e.g. prerequisite,\
          \ dependent on)"
        items:
          $ref: "#/definitions/ServiceOrderRelationship"
      relatedParty:
        type: "array"
        description: "A list of parties which are involved in this order and the role\
          \ they are playing"
        items:
          $ref: "#/definitions/RelatedParty"
      state:
        description: "State of the order: described in the state-machine diagram"
        $ref: "#/definitions/ServiceOrderStateType"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    example:
      note:
      - date: "2000-01-23T04:56:07.000+00:00"
        system: "system"
        '@baseType': "@baseType"
        author: "author"
        '@type': "@type"
        text: "text"
        '@schemaLocation': "http://example.com/aeiou"
      - date: "2000-01-23T04:56:07.000+00:00"
        system: "system"
        '@baseType': "@baseType"
        author: "author"
        '@type': "@type"
        text: "text"
        '@schemaLocation': "http://example.com/aeiou"
      orderItem:
      - '@baseType': "@baseType"
        service:
          serviceType: "serviceType"
          supportingResource:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          '@type': "@type"
          serviceSpecification:
            targetServiceSchema:
              '@baseType': "@baseType"
              '@type': "@type"
              '@schemaLocation': "@schemaLocation"
            '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
            version: "version"
          supportingService:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          relatedParty:
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          serviceRelationship:
          - relationshipType: "relationshipType"
            '@baseType': "@baseType"
            service:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
            '@type': "@type"
            '@schemaLocation': "http://example.com/aeiou"
          - relationshipType: "relationshipType"
            '@baseType': "@baseType"
            service:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
            '@type': "@type"
            '@schemaLocation': "http://example.com/aeiou"
          '@baseType': "@baseType"
          name: "name"
          id: "id"
          href: "href"
          place:
          - role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          state: {}
          category: "category"
          '@schemaLocation': "http://example.com/aeiou"
          serviceCharacteristic:
          - '@baseType': "@baseType"
            '@type': "@type"
            valueType: "valueType"
            name: "name"
            '@schemaLocation': "http://example.com/aeiou"
            value: {}
          - '@baseType': "@baseType"
            '@type': "@type"
            valueType: "valueType"
            name: "name"
            '@schemaLocation': "http://example.com/aeiou"
            value: {}
        '@type': "@type"
        action: {}
        appointment:
          '@referredType': "@referredType"
          '@baseType': "@baseType"
          '@type': "@type"
          description: "description"
          id: "id"
          href: "href"
          '@schemaLocation': "http://example.com/aeiou"
        orderItemRelationship:
        - relationshipType: "relationshipType"
          '@baseType': "@baseType"
          '@type': "@type"
          id: "id"
          '@schemaLocation': "http://example.com/aeiou"
        - relationshipType: "relationshipType"
          '@baseType': "@baseType"
          '@type': "@type"
          id: "id"
          '@schemaLocation': "http://example.com/aeiou"
        id: "id"
        state: {}
        '@schemaLocation': "http://example.com/aeiou"
      - '@baseType': "@baseType"
        service:
          serviceType: "serviceType"
          supportingResource:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          '@type': "@type"
          serviceSpecification:
            targetServiceSchema:
              '@baseType': "@baseType"
              '@type': "@type"
              '@schemaLocation': "@schemaLocation"
            '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
            version: "version"
          supportingService:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          relatedParty:
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          serviceRelationship:
          - relationshipType: "relationshipType"
            '@baseType': "@baseType"
            service:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
            '@type': "@type"
            '@schemaLocation': "http://example.com/aeiou"
          - relationshipType: "relationshipType"
            '@baseType': "@baseType"
            service:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
            '@type': "@type"
            '@schemaLocation': "http://example.com/aeiou"
          '@baseType': "@baseType"
          name: "name"
          id: "id"
          href: "href"
          place:
          - role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          state: {}
          category: "category"
          '@schemaLocation': "http://example.com/aeiou"
          serviceCharacteristic:
          - '@baseType': "@baseType"
            '@type': "@type"
            valueType: "valueType"
            name: "name"
            '@schemaLocation': "http://example.com/aeiou"
            value: {}
          - '@baseType': "@baseType"
            '@type': "@type"
            valueType: "valueType"
            name: "name"
            '@schemaLocation': "http://example.com/aeiou"
            value: {}
        '@type': "@type"
        action: {}
        appointment:
          '@referredType': "@referredType"
          '@baseType': "@baseType"
          '@type': "@type"
          description: "description"
          id: "id"
          href: "href"
          '@schemaLocation': "http://example.com/aeiou"
        orderItemRelationship:
        - relationshipType: "relationshipType"
          '@baseType': "@baseType"
          '@type': "@type"
          id: "id"
          '@schemaLocation': "http://example.com/aeiou"
        - relationshipType: "relationshipType"
          '@baseType': "@baseType"
          '@type': "@type"
          id: "id"
          '@schemaLocation': "http://example.com/aeiou"
        id: "id"
        state: {}
        '@schemaLocation': "http://example.com/aeiou"
      '@type': "@type"
      description: "description"
      externalId: "externalId"
      priority: "priority"
      relatedParty:
      - '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      - '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      requestedStartDate: "2000-01-23T04:56:07.000+00:00"
      '@baseType': "@baseType"
      notificationContact: "notificationContact"
      completionDate: "2000-01-23T04:56:07.000+00:00"
      expectedCompletionDate: "2000-01-23T04:56:07.000+00:00"
      id: "id"
      href: "href"
      category: "category"
      orderRelationship:
      - '@referredType': "@referredType"
        relationshipType: "relationshipType"
        '@baseType': "@baseType"
        '@type': "@type"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      - '@referredType': "@referredType"
        relationshipType: "relationshipType"
        '@baseType': "@baseType"
        '@type': "@type"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      '@schemaLocation': "http://example.com/aeiou"
      orderDate: "2000-01-23T04:56:07.000+00:00"
      startDate: "2000-01-23T04:56:07.000+00:00"
      requestedCompletionDate: "2000-01-23T04:56:07.000+00:00"
...